feat: implemeing the plot of the graphs (1,2,6) and descopleating the aux function logic into a aux directory
