+++ aux/auxiliar_functions.py
@@ -1,15 +1,34 @@
 import pandas as pd
 
+def map_times_by_rate(df_to_map):
+    times_dict = {}
+    for i in range(0, 11):
+        times_dict[i] = df_to_map.loc[df_to_map['rate'] == i, 'times_chosen'].values
+
+        if list(times_dict[i]) == []:
+            times_dict[i] = 0
+        else:
+            times_dict[i] = int(times_dict[i][0])
+
+    return times_dict
+
+def get_data_frame_app_rate(df):
+    df_rate_group = df["ia_confiavel"].value_counts().reset_index() # tranforma em um DataFrame (Series -> DataFrame)
+
+    df_rate_group.columns = ["rate", "times_chosen"]
+
+    return df_rate_group
+
 # faz o agrupamneto das especialidades demonstrando a quantidade total para cada especialidade
 # alem de destacar a porcentagem em relação a quntidade total de respostas
 def group_speciality_quantity(df):
-    df_group = df.groupby("Questão1")["ID"].count().rename("Times Chosen").reset_index()
+    df_group = df.groupby("area_atuacao")["data_hora"].count().rename("Times Chosen").reset_index()
 
     df_group_percent = add_speciality_percentage_column(df_group)
 
-    df_group_percent.set_index('Questão1', inplace=True)
+    df_group_percent.set_index('area_atuacao', inplace=True)
 
-    return df_group_percent.rename(columns={"Questão1": "Specialisties"})
+    return df_group_percent.rename(columns={"area_atuacao": "Specialisties"})
 
 # adiciona a coluna de porcentagem sob a quantidade
 # retorna um Data Frama que é agupado com quantidade e porcentagem total
@@ -29,11 +48,21 @@ def add_speciality_percentage_column(df_group):
 # Médico Residente             10         0          0                0               6
 # Técnico em Radiologia         1         0          0                0               1
 def group_percentage_years_specialist(df_orig):
-    result_percentage = pd.crosstab(df_orig["Questão1"], 
-                                    df_orig["Questão2"], 
+    result_percentage = pd.crosstab(df_orig["area_atuacao"], 
+                                    df_orig["experiencia"], 
                                     normalize='index') * 100 # realiza o calculo com base no total da linha
     
-    order_column = ["Menos de 1 ano", "1-3 anos", "4-7 anos", "8-15 anos", "Mais de 15 anos"] # ordenar as colunas
+    map_values_UI = {
+        'Não possuo experiência em laudos de exames de Radiologia': 'Nenhuma',
+    }
+    
+    result_percentage.columns = result_percentage.columns.map(lambda x: map_values_UI.get(x, x))
+    
+    order_column = ["Nenhuma", 
+                    "Menos de 1 ano", 
+                    "1-3 anos", "4-7 anos", 
+                    "8-15 anos", 
+                    "Mais de 15 anos"] # ordenar as colunas
 
     df_order_columns = result_percentage.reindex(columns=order_column).round(2) # a funcao 'reindex' re-indexa/substitui as colunas nessa ordem
 
@@ -41,17 +70,3 @@ def group_percentage_years_specialist(df_orig):
     df_order_columns.columns.name = None # remove a referencia à coluna original que referencia a coluna do DataFrame ao fazer o crosstab ("Questão2")
 
     return df_order_columns
-
-
-if __name__ == "__main__":
-    df = pd.read_csv("questionary_sheet.csv")
-
-    print("\n")
-    print("Dataframes sobre os anos para cada especialidade e sobre a quantidade, juntamente com a porcentagem: ")
-    print("\n============\n")
-
-    print(group_percentage_years_specialist(df))
-    print("\n============\n")
-
-    print(group_speciality_quantity(df))
-
