+++ main-script/README.md
@@ -2,11 +2,12 @@
 
 ## Demográficos
 
- - **1**: Proporção entre especialistas e residentes;
- - **2**: Tempo de experiência com laudos de Radiologia;
+ - **1**: Proporção entre cargos (João Neto);
+ - **2**: Tempo de experiência com laudos de Radiologia (João Neto);
 
 ## Demais
 
  - **3**: Confiança dos diagnósticos em casos de discordância e concordância da IA (e segunda opinião); (Guilherme)
  - **4**: Confiabilidade no uso de sistemas de IA para auxílio de decisão para interpretação de imagens (Escala 0-10) (Guilherme)
  - **5**: Funcionalidades que serviriam para aumentar a confiança em sistemas de IA na Radiologia (Guilherme);
+ - **6**: Avaliação dos radiologistas sobre a IA para o diagnóstico em exames do seto (João Neto);
